Split out controller manifest steps
This patch splits the overcloud_controller.pp and
overcloud_controller_pacemaker.pp manifests into
6 individual steps that we then re-construct
during deployment. This will allow us to inject
external per role data into the workflow. Eventually,
as we decompose the "controller" into separate roles
we should be able to eleminate the majority of the
old controller templates in total.

Depends-On: I2bdfc70d04a4fa14cefcf928ea5947fbe7529cf9
Change-Id: Ibd4f51f65da475bb20a6b08d7bda673f330a5464